refactor: API 서비스 함수 도메인별 분리
