+++ frontend/src/apis/services/space-content.service.ts
@@ -1,0 +1,15 @@
+import type {
+  CreateSpaceContentInput,
+  SpaceContent,
+} from '../../types/space-content.type';
+import { http } from '../http';
+
+export const spaceContentService = {
+  create: (data: CreateSpaceContentInput) =>
+    http.post<SpaceContent>('/space-contents', data),
+
+  getBySpaceId: (spaceId: number) =>
+    http.get<SpaceContent[]>(`/spaces/${spaceId}/contents`),
+
+  delete: (id: number) => http.delete<void>(`/space-contents/${id}`),
+};
