+++ frontend/src/apis/services/space.service.ts
@@ -1,0 +1,24 @@
+import type {
+  CreateSpaceInput,
+  Space,
+  SpaceWithContents,
+  UpdateSpaceInput,
+} from '../../types/space.type';
+import { http } from '../http';
+
+export const spaceService = {
+  getAll: (params?: { page?: number; limit?: number; search?: string }) =>
+    http.get<Space[]>('/spaces', params),
+
+  getById: (id: number) => http.get<Space>(`/spaces/${id}`),
+
+  getWithContents: (id: number) =>
+    http.get<SpaceWithContents>(`/spaces/${id}/contents`),
+
+  create: (data: CreateSpaceInput) => http.post<Space>('/spaces', data),
+
+  update: (id: number, data: UpdateSpaceInput) =>
+    http.patch<Space>(`/spaces/${id}`, data),
+
+  delete: (id: number) => http.delete<void>(`/spaces/${id}`),
+};
